refactor: config/, lib/ 제거, 기존 contants/ 오타 >> constants 로 변경
