+++ src/components/@features/Parties/PartiesMainStageCard.tsx
@@ -2,7 +2,7 @@
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import React from 'react';
-import { routes } from '@/config/routes';
+import { routes } from '@/constants/routes';
 
 const PartiesMainStageCard = () => {
   const router = useRouter();
