+++ src/components/@features/Parties/PartyRoomCard.tsx
@@ -2,7 +2,7 @@
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import React from 'react';
-import { routes } from '@/config/routes';
+import { routes } from '@/constants/routes';
 
 interface PartyRoomCardProps {
   value: number; // TODO: set proper type for value when api is connected
