+++ src/components/@features/Profile/ProfileSettings.tsx
@@ -6,8 +6,8 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import GoBackButton from '@/components/__legacy__/GoBackButton';
 import InputErrorMessage from '@/components/__legacy__/InputErrorMessage';
-import { cn } from '@/lib/utils';
 import { useProfileStore } from '@/store/profile';
+import { cn } from '@/utils/cn';
 
 const profileFormSchema = z.object({
   nickName: z
