+++ src/components/@shared/@atoms/Tag.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 import Typography from './Typography';
 
 interface CommonProps {
