+++ src/components/@shared/@atoms/TextButton.tsx
@@ -3,7 +3,7 @@ import type { ComponentProps, ReactNode } from 'react';
 import { forwardRef } from 'react';
 import Loading from '@/components/@shared/@atoms/Loading';
 import Typography from '@/components/@shared/@atoms/Typography';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 type TextButtonColor = 'basic' | 'primary' | 'secondary';
