+++ src/components/@shared/@atoms/Typography.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { forwardRef, ReactNode } from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 export type TypographyType =
   | 'title1'
