+++ src/components/__legacy__/Header.tsx
@@ -1,7 +1,7 @@
 'use client';
 import Link from 'next/link';
 import React from 'react';
-import { routes } from '@/config/routes';
+import { routes } from '@/constants/routes';
 import EmailBox from './EmailBox';
 import Icons from './Icons';
